rename rinsar to minsar
