+++ bashfiles/alias.bash
@@ -22,7 +22,7 @@ alias cdr='cd ${PARENTDIR}'
 alias cdrb='cd ${PARENTDIR}/bashfiles'
 alias cdrs='cd ${INT_SCR}'
 alias cdrsa='cd ${PARENTDIR}/samples'
-alias cdri='cd ${PARENTDIR}/rinsar'
+alias cdri='cd ${PARENTDIR}/minsar'
 alias cdrt='cd ${PARENTDIR}/sources/rsmas_tools'
 alias cdrts='cd ${PARENTDIR}/sources/rsmas_tools/SAR'
 alias cdrtg='cd ${PARENTDIR}/sources/rsmas_tools/GPS'
