+++ default_isce22.bash
@@ -51,7 +51,7 @@ export CONDA_ENVS_PATH=${PARENTDIR}/3rdparty/miniconda3/envs
 export CONDA_PREFIX=${PARENTDIR}/3rdparty/miniconda3
 export PROJ_LIB=${CONDA_PREFIX}/share/proj
 export GDAL_DATA=${PYTHON3DIR}/share/gdal
-export DASK_CONFIG=${RSMAS_INSAR}/rinsar/defaults/dask
+export DASK_CONFIG=${RSMAS_INSAR}/minsar/defaults/dask
 
 export PYTHONPATH=${PYTHONPATH-""}
 export PYTHONPATH=${PYTHONPATH}:${PYSAR_HOME}
@@ -74,8 +74,8 @@ export PYTHONWARNINGS="ignore:Unverified HTTPS request"
 export PATH=${PATH}:${SSARAHOME}
 export PATH=${PATH}:${SSARA_ASF}
 export PATH=${PATH}:${SQUEESAR}
-export PATH=${PATH}:${RSMAS_INSAR}/rinsar:${RSMAS_INSAR}/rinsar/utils
-export PATH=${PATH}:${PARENTDIR}/rinsar
+export PATH=${PATH}:${RSMAS_INSAR}/minsar:${RSMAS_INSAR}/minsar/utils
+export PATH=${PATH}:${PARENTDIR}/minsar
 export PATH=${PATH}:${PARENTDIR}/setup/accounts
 export PATH=${PATH}:${PARENTDIR}/sources/rsmas_tools/SAR:${PARENTDIR}/sources/rsmas_tools/GPS:${PARENTDIR}/sources/rsmas_tools/notebooks
 export PATH=${ISCE_HOME}/applications:${ISCE_HOME}/bin:${ISCE_STACK}:${PATH}
