+++ minsar/defaults/run_operations_LSF.job
