+++ minsar/defaults/stack_template.txt
