+++ minsar/docs/run_operations.md
