+++ minsar/run_operations.py
@@ -7,10 +7,10 @@ from datetime import datetime
 import shutil
 import time
 
-from rinsar.utils import generate_template_files
-from rinsar.objects.rsmas_logging import RsmasLogger, loglevel
-from rinsar.pbjects import dataset_template
-import rinsar.utils.process_utilities as putils
+from minsar.utils import generate_template_files
+from minsar.objects.rsmas_logging import RsmasLogger, loglevel
+from minsar.pbjects import dataset_template
+import minsar.utils.process_utilities as putils
 
 
 ############## DIRECTORY AND FILE CONSTANTS ##############
