+++ minsar/utils/examine_jobs.bash
