--- rinsar/defaults/template_defaults.cfg
@@ -1,63 +1,0 @@
-[processRsmas]
-process_flag                          = TRUE
-every_day_flag                        = yes
-login                                 = pegasus.ccs.miami.edu
-user                                  = None
-email_pysar                           = None
-email_insarmaps                       = None
-processingMethod                      = smallbaseline
-demMethod                             = bbox
-cleanopt                              = 0
-processor                             = isce
-hazard_products_flag                  = False
-
-[ssaraOptions]
-ssaraopt.platform                     = None
-ssaraopt.relativeOrbit                = None
-ssaraopt.frame                        = None
-ssaraopt.startDate                    = None
-ssaraopt.endDate                      = None
-ssaraopt.parallel                     = 10
-
-[topsStack]
-topsStack.slcDir                      = ./SLC
-topsStack.orbitDir                    = $SENTINEL_ORBITS
-topsStack.auxDir                      = $SENTINEL_AUX
-topsStack.workingDir                  = ./
-topsStack.demDir                      = None
-topsStack.master                      = None
-topsStack.numConnections              = 3
-topsStack.numOverlapConnections       = 3
-topsStack.subswath                    = 1 2 3
-topsStack.boundingBox                 = None
-topsStack.textCmd                     =
-topsStack.excludeDates                = None
-topsStack.includeDates                = None
-topsStack.azimuthLooks                = 3
-topsStack.rangeLooks                  = 9
-topsStack.filtStrength                = 0.5
-topsStack.esdCoherenceThreshold       = 0.85
-topsStack.snrMisregThreshold          = 10
-topsStack.unwMethod                   = snaphu
-topsStack.polarization                = vv
-topsStack.coregistration              = NESD
-topsStack.workflow                    = interferogram
-topsStack.startDate                   = None
-topsStack.stopDate                    = None
-topsStack.useGPU                      = False
-
-[squeesar]
-squeesar.plmethod                     = EMI
-squeesar.patch_size                   = 300
-squeesar.range_window                 = 21
-squeesar.azimuth_window               = 15
-squeesar.subset                       = None
-squeesar.dask_numWorkers              = 40
-
-
-
-
-
-
-
-
